Create README.md
myFlix-client 

myFlix is a client-side application built with react.js, react-bootstrap for styling, and using parcel as a bundler. The focus is building an application that communicates with a server-side REST Api to display movies from a database and register users. 

  

App Features 

login and registration with validation 

display a list of movies that are stored in redux stores 

Users can click and find the details about the movie 

Users can search for a movie 

Users can update their information 

Users can delete their accounts 

Users can add/remove a movie from their favorite movies list 

  

check out the live demo link:  

  

Technical Requirements 

The application uses a single-page application (SPA) 

The application uses state routing to navigate between views and share URLs 

The application provides the users the option to filter movies using a “search” feature 

The application uses Parcel as its build tool 

The application is written using the React library and in ES2015+ 

The application includes Bootstrap as a UI library for styling and responsiveness 

The application contains function components 

The application is hosted online 

  

  

dependencies 

    "bootstrap": "5.2.0", 

    "cors": "^2.8.5", 

    "prop-types": "^15.8.1", 

    "react": "^18.2.0", 

    "react-bootstrap": "2.5.0", 

    "react-dom": "^18.2.0", 

    "react-router-dom": "^6.9.0", 

    "react-scripts": "^5.0.1",